Squash Bug / Allow More Pumps
Fix an issue which might cause drink dispensing to fail.  
Rewrite of certain parts of the app.py, control.py and common.py to allow for more pumps if needed.  If you wish to add more pumps, simply modify the settings.json to define more "pump_xx" definitions.  You should be able to support as many pumps as there are GPIO pins.